requirements.txt
Flask==3.1.2
pymongo==4.15.3
dnspython==2.8.0
python-dotenv==1.0.1